+++ examples/websocket_api/market/TickerPrice/TickerPrice.go
@@ -13,7 +13,7 @@ func main() {
 }
 
 func TickerPrice() {
-	client := binance_connector.NewWebsocketAPIClient("", "", "wss://testnet.binance.vision/ws-api/v3")
+	client := binance_connector.NewWebsocketAPIClient("", "", "wss://ws-api.testnet.binance.vision/ws-api/v3")
 	err := client.Connect()
 	if err != nil {
 		log.Printf("Error: %v", err)