+++ examples/websocket_api/trading/CurrentOpenOCO/CurrentOpenOCO.go
@@ -13,7 +13,7 @@ func main() {
 }
 
 func CurrentOpenOCOExample() {
-	client := binance_connector.NewWebsocketAPIClient("api_key", "secret_key", "wss://testnet.binance.vision/ws-api/v3")
+	client := binance_connector.NewWebsocketAPIClient("api_key", "secret_key", "wss://ws-api.testnet.binance.vision/ws-api/v3")
 	err := client.Connect()
 	if err != nil {
 		log.Printf("Error: %v", err)